+++ client/src/actions/index.js
@@ -7,6 +7,23 @@ export const moveToken = (tokenId, x, y) => {
 	}
 }
 
+export const rollDice = (dType, dNum, rolls, total) => {
+	return {
+		type: 'ROLL_DICE',
+		dType: dType,
+		dNum: dNum,
+		rolls: rolls,
+		total: total,
+	}
+}
+
+export const sendChatMessage = (text) => {
+	return {
+		type: 'CHAT_MESSAGE',
+		text: text,
+	}
+}
+
 
 //----- local actions -----
 export const setZoom = (viewScale) => {
@@ -29,22 +46,3 @@ export const closeTool = (tool) => {
 		tool: tool,
 	}
 }
-
-
-export const rollDice = (dType, dNum, rolls, total) => {
-	return {
-		type: 'ROLL_DICE',
-		dType: dType,
-		dNum: dNum,
-		rolls: rolls,
-		total: total,
-	}
-}
-
-export const sendChatMessage = (text) => {
-	return {
-		type: 'CHAT_MESSAGE',
-		text: text,
-	}
-}
-