+++ client/src/components/LoginPage.js
@@ -1,88 +1,79 @@
 import React from 'react';
+import { connect } from "react-redux";
 
-import { UserService } from '../services/UserService';
+import { login, logout } from '../actions/userActions';
 
 
-export default class LoginPage extends React.Component {
+class LoginPage extends React.Component {
 	constructor(props) {
 		super(props);
 
-		UserService.logout();
-
 		this.state = {
-			username: '',
-			password: '',
-			submitted: false,
-			loading: false,
-			error: ''
+			username: "",
+			password: "",
 		};
 
-		this.handleChange = this.handleChange.bind(this);
-		this.handleSubmit = this.handleSubmit.bind(this);
+		this.props.logout();
 	}
 
-	handleChange(e) {
+	inputHandler(e) {
 		const { name, value } = e.target;
 		this.setState({ [name]: value });
 	}
 
-	handleSubmit(e) {
+	submitHandler(e) {
 		e.preventDefault();
 
-		this.setState({ submitted: true });
-		//const { username, password, returnUrl } = this.state;
 		const { username, password } = this.state;
 
-		// stop here if form is invalid
-		if (!(username && password)) {
+		if ((!username) || (username === "") || (!password) || (password === "")) {
 			return;
 		}
 
-		this.setState({ loading: true });
-		UserService.login(username, password)
-			.then(
-				user => {
-					const { from } = this.props.location.state || { from: { pathname: "/" } };
-					this.props.history.push(from);
-				},
-				error => this.setState({ error, loading: false })
-			);
+		let targetUrl;
+		if (this.props.location.state)
+			targetUrl = this.props.location.state.from.pathName;
+		else
+			targetUrl = "/";
+
+		this.props.login(username, password, targetUrl);
 	}
 
 
 	render() {
-		const { username, password, submitted, loading, error } = this.state;
 		return (
 			<div className="">
 				<h2>Login</h2>
-				<form name="form" onSubmit={this.handleSubmit}>
-					<div className={'form-group' + (submitted && !username ? ' has-error' : '')}>
+				<form onSubmit={this.submitHandler.bind(this)}>
+					<div>
 						<label htmlFor="username">Username</label>
-						<input type="text" className="form-control" name="username" value={username} onChange={this.handleChange} />
-						{submitted && !username &&
-							<div className="help-block">Username is required</div>
-						}
+						<input type="text" name="username" value={this.state.username} 
+							onChange={this.inputHandler.bind(this)} />
 					</div>
-					<div className={'form-group' + (submitted && !password ? ' has-error' : '')}>
+					<div>
 						<label htmlFor="password">Password</label>
-						<input type="password" className="form-control" name="password" value={password} onChange={this.handleChange} />
-						{submitted && !password &&
-							<div className="help-block">Password is required</div>
-						}
+						<input type="password" name="password" value={this.state.password}
+							onChange={this.inputHandler.bind(this)} />
 					</div>
-					<div className="form-group">
-						<button className="btn btn-primary" disabled={loading}>Login</button>
+					<div>
+						<button disabled={this.props.loading}>Login</button>
 						{
 							null
 							//TODO: loading spinner
 							//loading && <img src="" />
 						}
 					</div>
-					{error &&
-						<div className={'alert alert-danger'}>{error}</div>
+					{ this.props.error &&
+						<div>{this.props.error}</div>
 					}
 				</form>
 			</div>
 		);
 	}
 }
+
+
+export default connect(
+	state => ({}),
+	{ login, logout }
+)(LoginPage);